requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
boto3==1.27.0
botocore==1.30.0
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==3.3.2
cryptography==41.0.2
Django==4.2.2
django-bootstrap-form==3.4
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.6
install==1.3.5
jmespath==1.0.1
mysqlclient==2.1.1
paramiko==3.2.0
Pillow==9.5.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
storage==0.0.4.3
tzdata==2023.3
urllib3==1.26.16
xlwt==1.3.0
gunicorn==20.1.0
psycopg2==2.8.6
whitenoise==6.2.0
